+++ source/Databricks/documents/release-notes/release-notes.md
@@ -1,5 +1,10 @@
 # Databricks Release Notes
 
+## Version 8.0.0
+
+- Obsolete methods are removed for Databricks SQL Warehouse integration
+- Add support for adhoc SQL queries
+
 ## Version 7.2.2
 
 - Databricks integration test configuration keys in keyvault are renamed